Update library for 64 bit
Conflicts:
	Airship/Common/UAAnalytics+Internal.h
	Airship/Common/UASQLite.m
	Airship/External/UA_fmdb/UA_FMDatabase.m
	Airship/Inbox/UAInboxMessage.m
	Airship/Inbox/UAInboxMessageList.m
	Airship/Inbox/UAInboxURLCache.m

+++ Airship/Common/UAUser.m
@@ -254,7 +254,7 @@ NSString * const UAUserCreatedNotification = @"com.urbanairship.notification.use
     [self.apiClient updateDeviceToken:deviceToken forUsername:self.username onSuccess:^(NSString *updatedToken) {
         UA_LINFO(@"Device token updated to: %@", updatedToken);
     } onFailure:^(UAHTTPRequest *request) {
-        UA_LDEBUG(@"Device token update failed with status: %d", request.response.statusCode);
+        UA_LDEBUG(@"Device token update failed with status: %ld", (long)request.response.statusCode);
     }];
 }
